again with the PDF interfaces
