+++ src/main/java/com/mindee/input/InputSourceUtils.java
@@ -1,12 +1,6 @@
 package com.mindee.input;
 
 import com.mindee.MindeeException;
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import org.apache.pdfbox.Loader;
-import org.apache.pdfbox.io.RandomAccessReadBuffer;
-import org.apache.pdfbox.pdmodel.PDDocument;
-import org.apache.pdfbox.text.PDFTextStripper;
 
 /**
  * Utilities for working with files.
@@ -64,47 +58,5 @@ public class InputSourceUtils {
       throw new MindeeException("File name must include a valid extension.");
     }
     return new String[] { name, extension };
-  }
-
-  /**
-   * Returns true if the file is a PDF.
-   */
-  public static boolean isPdf(byte[] fileBytes) {
-    try {
-      Loader.loadPDF(new RandomAccessReadBuffer(new ByteArrayInputStream(fileBytes)));
-    } catch (IOException e) {
-      return false;
-    }
-    return true;
-  }
-
-  /**
-   * Returns true if the source PDF has source text inside. Returns false for images.
-   *
-   * @param fileBytes A byte array representing a PDF.
-   * @return True if at least one character exists in one page.
-   * @throws MindeeException if the file could not be read.
-   */
-  public static boolean hasSourceText(byte[] fileBytes) {
-    try {
-      PDDocument document = Loader
-        .loadPDF(new RandomAccessReadBuffer(new ByteArrayInputStream(fileBytes)));
-      PDFTextStripper stripper = new PDFTextStripper();
-
-      for (int i = 0; i < document.getNumberOfPages(); i++) {
-        stripper.setStartPage(i + 1);
-        stripper.setEndPage(i + 1);
-        String pageText = stripper.getText(document);
-        if (!pageText.trim().isEmpty()) {
-          document.close();
-          return true;
-        }
-      }
-      document.close();
-    } catch (IOException e) {
-      return false;
-    }
-
-    return false;
   }
 }
